// src/enums/sector.ts
export enum Sector {
    ENERGY = 'ENERGY',
    MATERIALS = 'MATERIALS',
    INDUSTRIALS = 'INDUSTRIALS',
    UTILITIES = 'UTILITIES',
    HEALTHCARE = 'HEALTHCARE',
    FINANCIALS = 'FINANCIALS',
    CONSUMER_DISCRETIONARY = 'CONSUMER_DISCRETIONARY',
    CONSUMER_STAPLES = 'CONSUMER_STAPLES',
    INFORMATION_TECHNOLOGY = 'INFORMATION_TECHNOLOGY',
    COMMUNICATION = 'COMMUNICATION',
    REAL_ESTATE = 'REAL_ESTATE',
    ETF = 'ETF',
    INDEX_FUNDS = 'INDEX_FUNDS',
}